requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
mysql-connector-python==8.3.0
Pillow==10.3.0